feat: LocationManager 타입 구현.
- 기존 뷰컨에서 확인했던 기능들, 위의 타입의 기능으로 이동.
