+++ WeatherForecast/WeatherForecast/ViewController.swift
@@ -5,36 +5,12 @@
 // 
 
 import UIKit
-import CoreLocation
+
 
 class ViewController: UIViewController {
-    var locationManager = CLLocationManager()
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        setupLocationInfo()
-    }
-    
-    func setupLocationInfo() {
-        locationManager.delegate = self
-        locationManager.requestWhenInUseAuthorization()
-        locationManager.startUpdatingLocation()
     }
 }
 
-extension ViewController: CLLocationManagerDelegate {
-    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
-        guard let currentLocation = locations.last else {
-            return
-        }
-        
-        let longitude = currentLocation.coordinate.longitude
-        let latitude = currentLocation.coordinate.latitude
-        
-        locationManager.stopUpdatingLocation()
-    }
-    
-    func locationManager(_ manager: CLLocationManager, didFailWithError error: Error) {
-        
-    }
-}
